pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = "an e-commerce api with fastapi"
authors = ["Felix <felix.landlord.nunoo@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
tortoise-orm = "^0.21.6"
fastapi = {extras = ["standard"], version = "^0.115.0"}
uvicorn = "^0.31.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"